Move OBJ multicluster access key tests from `objectStorage` to `objectStorageMulticluster` dir
